✨ feat : Add header logo refresh and current time display
- Implemented page refresh on header logo click
- Displayed current time in header using YYYY.MM.DD (Day) format
- Created utility function to format date as specified
